refactor: componentes LISTMYGROUPS e LISTOTHERGROUPS foram unificados e a renderização de grupos agora é modificada por variáveis, onde o componente LISTMYGROUPS renderiza ambas listas a depender do parâmetro. O componente LISTOTHERGROUPS foi excluído
