+++ src/components/ListMyGroups/ListMyGroups.module.css
@@ -1,0 +1,15 @@
+.list_my_groups_container {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    width: 100%; 
+    margin-bottom: 6rem;
+}
+
+.list_other_groups_container {
+    display: flex;
+    flex-wrap: wrap;
+    height: 100%;
+    width: 100%;
+    margin-bottom: 6rem;
+}
