--- src/components/ListOtherGroups/ListOtherGroups.module.css
@@ -1,7 +1,0 @@
-.list_container {
-    display: flex;
-    /* flex-direction: column; */
-    flex-wrap: wrap;
-    height: 100%;
-    width: 100%; 
-}
